// Java/Tamin-Projects/eservice-lite/ui/src/app/models/registration/relationTamin.model.ts
export class RelationTaminModel {
     insuranceId: string;
        parentId: string;
        parentIsPensioner: string;
        nationalId: string;
        firstName: string;
        lastName: string;
        organizationName: string;
        birthDate: string;
        workshopHelp: string;
        startDate: string;
        endDate: string;
        workshopId: string;
        workshopName: string;
        paymentSequence: string;
        pensionerId: string;
        unemploymentAmount: string;
        workshopJob: string;
        relationWithTamin: any;
        dependentTypeId: string;
        ageFlag: string;
        mohlatDate: string;
}